requirements.txt
Flask==3.1.1
psycopg2-binary==2.9.7
supabase==2.3.3
python-dotenv==1.0.0
bcrypt==4.1.2
cryptography==42.0.5 # Añadimos esta línea